feat: Support local charts on file system
This adds support for charts that are on the file system rather than in a repo. It works by checking
whether a chart has a repo that's listed in the repos section of the DSF. If it doesn't, the chart
name is treated as a file path instead. If the chart is not present on the file system, or the chart
on the file system is a different version than specified in the DSF, validation fails.