Fix bug with diff-context flag
diffRelease() expects its flags to end with whitespace, which
diffContextFlag does not. This bug prevents the flag from working at
all.

A less brittle approach generally would be to pass all flags as separate
elements of the Args field, leaving them to be joined by cmd.exec().
This would however result in variable numbers of spaces depending on
which flags are enabled.